refac: 코드 최적화 진행

+++ src/main/java/com/example/sns/service/PostService.java
@@ -47,8 +47,12 @@ public class PostService {
 
     @Transactional
     public void delete(Integer userId, Integer postId) {
-        PostEntity postEntity = getPostEntityorException(postId);
-        isSameId(userId, postId, postEntity);
+        PostEntity postEntity = postEntityRepository.findById(postId).orElseThrow(() -> new SimpleSnsApplicationException(ErrorCode.POST_NOT_FOUND, String.format("postId is %d", postId)));
+        if (!Objects.equals(postEntity.getUser().getId(), userId)) {
+            throw new SimpleSnsApplicationException(ErrorCode.INVALID_PERMISSION, String.format("user %s has no permission with post %d", userId, postId));
+        }
+        likeEntityRepository.deleteAllByPost(postEntity);
+        commentEntityRepository.deleteAllByPost(postEntity);
         postEntityRepository.delete(postEntity);
     }
